extract SlowQuerieTable
