+++ ui/lib/apps/SlowQuery/components/List.tsx
@@ -1,20 +1,16 @@
 import React, { useState, useEffect } from 'react'
-import { useNavigate } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import { Select, Space, Tooltip, Input } from 'antd'
 import { ReloadOutlined } from '@ant-design/icons'
 import { ScrollablePane } from 'office-ui-fabric-react/lib/ScrollablePane'
-import { IColumn } from 'office-ui-fabric-react/lib/DetailsList'
 import { useSessionStorageState } from '@umijs/hooks'
-import { Card, CardTableV2 } from '@lib/components'
+import { Card } from '@lib/components'
 import client, { SlowqueryBase } from '@lib/client'
-import * as useColumn from '@lib/utils/useColumn'
 import TimeRangeSelector, {
   TimeRange,
   DEF_TIME_RANGE,
 } from './TimeRangeSelector'
-import DetailPage from './Detail'
-import * as useSlowQueryColumn from '../utils/useColumn'
+import SlowQueriesTable, { OrderBy } from './SlowQueriesTable'
 
 import styles from './List.module.less'
 
@@ -23,8 +19,6 @@ const { Search } = Input
 
 const SEARCH_OPTIONS_SESSION_KEY = 'slow_query_search_options'
 const LIMITS = [100, 200, 500, 1000]
-
-type OrderBy = 'Query_time' | 'Mem_max' | 'Time'
 
 export interface ISearchOptions {
   timeRange: TimeRange
@@ -44,38 +38,7 @@ const defSearchOptions: ISearchOptions = {
   limit: 100,
 }
 
-function tableColumns(
-  rows: SlowqueryBase[],
-  onColumnClick: (ev: React.MouseEvent<HTMLElement>, column: IColumn) => void,
-  orderBy: OrderBy,
-  desc: boolean
-): IColumn[] {
-  return [
-    useSlowQueryColumn.useSqlColumn(rows),
-    {
-      ...useSlowQueryColumn.useTimestampColumn(rows),
-      isSorted: orderBy === 'Time',
-      isSortedDescending: desc,
-      onColumnClick: onColumnClick,
-    },
-    {
-      ...useSlowQueryColumn.useQueryTimeColumn(rows),
-      isSorted: orderBy === 'Query_time',
-      isSortedDescending: desc,
-      onColumnClick: onColumnClick,
-    },
-    {
-      ...useSlowQueryColumn.useMemoryColumn(rows),
-      isSorted: orderBy === 'Mem_max',
-      isSortedDescending: desc,
-      onColumnClick: onColumnClick,
-    },
-    useColumn.useDummyColumn(),
-  ]
-}
-
 function List() {
-  const navigate = useNavigate()
   const { t } = useTranslation()
 
   const [searchOptions, setSearchOptions] = useSessionStorageState(
@@ -87,13 +50,6 @@ function List() {
   const [refreshTimes, setRefreshTimes] = useState(0)
   const [allSchemas, setAllSchemas] = useState<string[]>([])
   const [slowQueryList, setSlowQueryList] = useState<SlowqueryBase[]>([])
-
-  const columns = tableColumns(
-    slowQueryList || [],
-    onColumnClick,
-    searchOptions.orderBy,
-    searchOptions.desc
-  )
 
   useEffect(() => {
     async function getSchemas() {
@@ -122,30 +78,6 @@ function List() {
     }
     getSlowQueryList()
   }, [searchOptions, refreshTimes])
-
-  function onColumnClick(_ev: React.MouseEvent<HTMLElement>, column: IColumn) {
-    if (column.key === searchOptions.orderBy) {
-      setSearchOptions({
-        ...searchOptions,
-        desc: !searchOptions.desc,
-      })
-    } else {
-      setSearchOptions({
-        ...searchOptions,
-        orderBy: column.key as OrderBy,
-        desc: true,
-      })
-    }
-  }
-
-  function handleRowClick(rec) {
-    const qs = DetailPage.buildQuery({
-      digest: rec.digest,
-      connectId: rec.connection_id,
-      time: rec.timestamp,
-    })
-    navigate(`/slow_query/detail?${qs}`)
-  }
 
   return (
     <ScrollablePane style={{ height: '100vh' }}>
@@ -203,11 +135,18 @@ function List() {
           </Space>
         </div>
       </Card>
-      <CardTableV2
+      <SlowQueriesTable
         loading={loading}
-        items={slowQueryList || []}
-        columns={columns}
-        onRowClicked={handleRowClick}
+        slowQueries={slowQueryList}
+        orderBy={searchOptions.orderBy}
+        desc={searchOptions.desc}
+        onChangeSort={(orderBy, desc) =>
+          setSearchOptions({
+            ...searchOptions,
+            orderBy,
+            desc,
+          })
+        }
       />
     </ScrollablePane>
   )
